requirements.txt
mcp>=1.0.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
SQLAlchemy>=2.0.0
prefab-ui>=0.19.0
altair>=5.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
psutil>=5.9.0
streamlit>=1.30.0
google-adk>=0.1.0
google-genai>=0.1.0
pandas>=2.0.0
